added a GPA calculator
at the moment, it doesn't calculate the GPA, but makes the distinction as to whether or not you're on the deans, presidents, or academic danger list. In 2014, I came very close to being put in the academic danger list because my English teacher waited until the last minute to enter our grades.