added debug prompter
When I'm debugging a program, it displays the input and lets me know if it exits. However, should a program like this be run in a production environment, I won't see that prompt and the program will open and close too quickly for me to see what happened.

So I created this function and added it to my assignments so that anyone who reads my programs won't watch it go that fast. I also address the 'press any key' running gag with a common sense addendum. This function needs a function prototype in order to work, and a function called program() as calling main() isn't exactly reasonable.